Allow js-with-ts cli to take filepath as arg and output .ts
The original implementation took input from stdin and sent output to
stdout only. However, it's useful to be able to walk a directory of
javascript files and drop a .ts file wherever we find a .js file.

Rather than make that file pairing a task of the build system, this
commit makes it an easy command-line arg.

For example:

$ js-with-ts index.js

Will output a companion file, index.ts