Add types to core package
During most development, we don't need a build process because all
source files are just javascript files. However, it turns out there
is a special case for when we want to bundle code up as an NPM
package, and we want to include types, we need a build step.

This commit adds the "built" types to the core package.

Note: if no types are needed, then no build step is necessary,
  because an NPM package with src files as JS files is perfect.