Allow importing TS extensions
In almost all cases, where js-with-ts is involved, we'd like to be
able to import types from ".ts" files that accompany the project's
".js" files.

Note that we ONLY want ".ts" files to contain type information, not
code, so that we can import the types via js-with-ts comments.

Example usage:

/*+ import type { Item } from "./types.ts" */