Import type must go inside TS block
Although adding an import like this to a JS file does not cause a
type error, node will fail to parse/import because it can't find the
".ts" file.

Currently, this is a shortcoming of the js-with-ts VSCode extension,
as it treats most of the file as TS, not "JS interleaved with TS",
which would be ideal.