No longer require the mount/head function to return builtins
Previously, when modifying the <head /> in any way, it was required that
the `mount({ head() { ... } })` (head) function returned the builtins it
was passed. This was not very ergonomic, because it required you to know
the internals of the builtins-creation process.

Now, the `head` function behaves more like the `body` function: you just
return html as needed, and it will be appended to the head section of the
HTML. In rare cases where you need to modify the builtin inclusions that
are required by Credit, there is a `builtins(m)` function that allows you
to pass a mapper function. The default mapper is the identity function:

function mapper(builtins) {
  return builtins
}

But you could modify it by adding any html before or after the builtins,
(or removing the builtins altogether):

function mapper (builtins) {
  return html`${builtins}\n<script>...</script>`
}

Note that this will only affect the server-side rendering.