Experiment with different importmaps per environment
The importmap for development depends on a server running at port 9000
locally, so that the "latest" @credit/core code can be loaded by the
browser.

An alternative importmap for production depends on jsdelivr servers on
the open internet. This may be convenient for most people, especially
if they do not want to experiment with or tweak the core.

Currently, @credit/core has not been published, so the production env
will not work. Hopefully this will change soon.