Rename js-with-ts to ts-dot-js, and require .ts.js file extension
The js-with-ts packages have been renamed to ts-dot-js to reflect the
requirement that files end in ".ts.js" (this may become configurable in
the future).

The motivation for this change was the fact that the vscode extension
was ambiently modifying the javascript language context--meaning that
once installed, the extension just assumed all JS files were to support
typescript within block comments.

This was not sustainable, as it is often the case that some projects or
subdirs have regular JS files, while other projects may have JS with
TS comments.

Now, you can select a specific language mode ("TS dot JS") when opening
a file, and `.ts.js` files in general are associated with this modified
language mode.