Check for "ts-dot-js" customConditions in tsconfig.json
Sometimes, we need a plain .js file to be treated as a .ts.js file. For
example, when compiling types for npm packaging using tsc.

Rather than hard-coding ".ts.js" as the only way for the typescript-plugin
to determine whether a file conforms to the "ts-dot-js" comment spec,
allow for the tsconfig.json file in a project to also specify that files
should be parsed this way.

The special condition that allows this is the "customConditions" clause
of the tsconfig.json file. Its values should include: ["ts-dot-js"].