Switch to typescript language server for ts-dot-js files
The typescript language server is essential to making the ts-dot-js hack
work. However, because we use the ".js" file ending, vscode does not
automatically recognize that the "typescript" language should be chosen.

Use the vscode extension to switch to typescript language when we have
a `tsconfig.json` file configured for ts-dot-js:

```
{
  "compilerOptions": {
    "customConditions": ["ts-dot-js"]
  },
}
```